Fixed require cache object persistence
When you had the field .persist on false it would get removed from the json as intended during the lifespan of the insert creation. However this was not the case and the field got removed and cached as a result of how require() works.

This has been reworked to create a temp object which will filter on the original json and keep the original json as a 'read_only' which is safer.